Drastically reduced memory consumption and fixed leak

+++ src/Chunk.h
@@ -21,7 +21,7 @@ typedef struct {
 
 typedef struct {
   ChunkMesh *mesh;
-  uint32_t *map; // This is 3D
+  uint8_t *map; // This is 3D
   bool is_updated;
   bool is_prepared;
 } Chunk;
@@ -30,7 +30,7 @@ ChunkMesh *chunckmesh_init();
 
 // Chunk* chunk_build(uint32_t** map, uint32_t width,uint32_t height,uint32_t
 // depth);
-Chunk *chunk_build(uint32_t map[CHUNK_DEPTH * CHUNK_WIDTH * CHUNK_HEIGHT]);
+Chunk *chunk_build(uint8_t map[CHUNK_DEPTH * CHUNK_WIDTH * CHUNK_HEIGHT]);
 void chunk_destroy(Chunk **chunk);
 void chunk_prepare(Chunk *chunk);
 // void chunk_update(Chunk* chunk);
