+++ src/worldgen.h
@@ -2,5 +2,5 @@
 #define BVCRAFT_WORLDGEN_H
 
 #include "Block.h"
-uint32_t * worldgen_random(uint32_t width, uint32_t height, uint32_t depth);
+uint8_t * worldgen_random(uint32_t width, uint32_t height, uint32_t depth);
 #endif
